Revert "chore(fleet): Add APM SSI status to inventoryagent payload (#36875)"
This reverts commit af012956fe2b0632e6a74fa9a28c80d9be6752a6.
