+++ comp/README.md
@@ -646,12 +646,6 @@ Package status implements the core status component information provider interfa
 
 Package updater implements the updater component.
 
-### [comp/updater/installerexec](https://pkg.go.dev/github.com/DataDog/datadog-agent/comp/updater/installerexec)
-
-*Datadog Team*: fleet
-
-Package installerexec provides a component to execute installer commands
-
 ### [comp/updater/localapi](https://pkg.go.dev/github.com/DataDog/datadog-agent/comp/updater/localapi)
 
 Package localapi is the updater local api component.
@@ -659,13 +653,6 @@ Package localapi is the updater local api component.
 ### [comp/updater/localapiclient](https://pkg.go.dev/github.com/DataDog/datadog-agent/comp/updater/localapiclient)
 
 Package localapiclient provides the local API client component.
-
-### [comp/updater/ssistatus](https://pkg.go.dev/github.com/DataDog/datadog-agent/comp/updater/ssistatus)
-
-*Datadog Team*: fleet
-
-Package ssistatus is a component to regularly retrieve the status of APM Single Step Instrumentation and
-add it to the inventoryagent payload.
 
 ### [comp/updater/telemetry](https://pkg.go.dev/github.com/DataDog/datadog-agent/comp/updater/telemetry)
