--- comp/updater/ssistatus/def/component.go
@@ -1,13 +1,0 @@
-// Unless explicitly stated otherwise all files in this repository are licensed
-// under the Apache License Version 2.0.
-// This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2025-present Datadog, Inc.
-
-// Package ssistatus is a component to regularly retrieve the status of APM Single Step Instrumentation and
-// add it to the inventoryagent payload.
-package ssistatus
-
-// team: fleet
-
-// Component is the component type.
-type Component interface{}
