+++ pkg/fleet/installer/commands/command.go
@@ -25,7 +25,6 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/fleet/installer/repository"
 	"github.com/DataDog/datadog-agent/pkg/fleet/installer/setup"
 	"github.com/DataDog/datadog-agent/pkg/fleet/installer/telemetry"
-	installertypes "github.com/DataDog/datadog-agent/pkg/fleet/installer/types"
 	"github.com/DataDog/datadog-agent/pkg/version"
 )
 
@@ -80,7 +79,7 @@ func (c *cmd) stop(err error) {
 
 type installerCmd struct {
 	*cmd
-	installertypes.Installer
+	installer.Installer
 }
 
 func newInstallerCmd(operation string) (_ *installerCmd, err error) {
@@ -90,7 +89,7 @@ func newInstallerCmd(operation string) (_ *installerCmd, err error) {
 			cmd.stop(err)
 		}
 	}()
-	var i installertypes.Installer
+	var i installer.Installer
 	if MockInstaller != nil {
 		i = MockInstaller
 	} else {
@@ -166,6 +165,7 @@ func RootCommands() []*cobra.Command {
 		promoteConfigExperimentCommand(),
 		garbageCollectCommand(),
 		purgeCommand(),
+		isInstalledCommand(),
 		apmCommands(),
 		getStateCommand(),
 		statusCommand(),
@@ -181,7 +181,6 @@ func UnprivilegedCommands() []*cobra.Command {
 	return []*cobra.Command{
 		versionCommand(),
 		defaultPackagesCommand(),
-		isInstalledCommand(),
 	}
 }
