+++ pkg/fleet/installer/commands/status.go
@@ -9,6 +9,7 @@ import (
 	"bytes"
 	_ "embed"
 	"encoding/json"
+	"errors"
 	"fmt"
 	"os"
 	"os/exec"
@@ -16,7 +17,6 @@ import (
 	"github.com/fatih/color"
 	"github.com/spf13/cobra"
 
-	"github.com/DataDog/datadog-agent/pkg/fleet/installer/packages/ssi"
 	"github.com/DataDog/datadog-agent/pkg/fleet/installer/repository"
 	template "github.com/DataDog/datadog-agent/pkg/template/html"
 	"github.com/DataDog/datadog-agent/pkg/version"
@@ -55,10 +55,17 @@ var functions = template.FuncMap{
 }
 
 type statusResponse struct {
-	Version            string                       `json:"version"`
-	Packages           *repository.PackageStates    `json:"packages"`
-	ApmInjectionStatus ssi.APMInstrumentationStatus `json:"apm_injection_status"`
-	RemoteConfigState  []*remoteConfigPackageState  `json:"remote_config_state"`
+	Version            string                      `json:"version"`
+	Packages           *repository.PackageStates   `json:"packages"`
+	ApmInjectionStatus apmInjectionStatus          `json:"apm_injection_status"`
+	RemoteConfigState  []*remoteConfigPackageState `json:"remote_config_state"`
+}
+
+// apmInjectionStatus contains the instrumentation status of the APM injection.
+type apmInjectionStatus struct {
+	HostInstrumented   bool `json:"host_instrumented"`
+	DockerInstalled    bool `json:"docker_installed"`
+	DockerInstrumented bool `json:"docker_instrumented"`
 }
 
 func status(debug bool, jsonOutput bool) error {
@@ -73,7 +80,7 @@ func status(debug bool, jsonOutput bool) error {
 		return fmt.Errorf("error getting package states: %w", err)
 	}
 
-	apmSSIStatus, err := ssi.GetInstrumentationStatus()
+	apmSSIStatus, err := getAPMInjectionStatus()
 	if err != nil {
 		fmt.Fprintf(os.Stderr, "error getting APM injection status: %s", err.Error())
 	}
@@ -106,6 +113,41 @@ func status(debug bool, jsonOutput bool) error {
 		fmt.Println(string(rawResult))
 	}
 	return nil
+}
+
+func getAPMInjectionStatus() (status apmInjectionStatus, err error) {
+	// Host is instrumented if the ld.so.preload file contains the apm injector
+	ldPreloadContent, err := os.ReadFile("/etc/ld.so.preload")
+	if err != nil && !errors.Is(err, os.ErrNotExist) {
+		return status, fmt.Errorf("could not read /etc/ld.so.preload: %w", err)
+	}
+	if bytes.Contains(ldPreloadContent, []byte("/opt/datadog-packages/datadog-apm-inject/stable/inject")) {
+		status.HostInstrumented = true
+	}
+
+	// Docker is installed if the docker binary is in the PATH
+	_, err = exec.LookPath("docker")
+	if err != nil && errors.Is(err, exec.ErrNotFound) {
+		return status, nil
+	} else if err != nil {
+		return status, fmt.Errorf("could not check if docker is installed: %w", err)
+	}
+	status.DockerInstalled = true
+
+	// Docker is instrumented if there is the injector runtime in its configuration
+	// We're not retrieving the default runtime from the docker daemon as we are not
+	// root
+	dockerConfigContent, err := os.ReadFile("/etc/docker/daemon.json")
+	if err != nil && !errors.Is(err, os.ErrNotExist) {
+		return status, fmt.Errorf("could not read /etc/docker/daemon.json: %w", err)
+	} else if errors.Is(err, os.ErrNotExist) {
+		return status, nil
+	}
+	if bytes.Contains(dockerConfigContent, []byte("/opt/datadog-packages/datadog-apm-inject/stable/inject")) {
+		status.DockerInstrumented = true
+	}
+
+	return status, nil
 }
 
 // remoteConfigState is the response to the daemon status route.
