+++ pkg/fleet/installer/installer.go
@@ -12,7 +12,6 @@ import (
 	"errors"
 	"fmt"
 	"os"
-	"os/user"
 	"path/filepath"
 	"runtime"
 	"strings"
@@ -30,7 +29,6 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/fleet/installer/oci"
 	"github.com/DataDog/datadog-agent/pkg/fleet/installer/packages"
 	"github.com/DataDog/datadog-agent/pkg/fleet/installer/repository"
-	installertypes "github.com/DataDog/datadog-agent/pkg/fleet/installer/types"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 	"github.com/DataDog/datadog-agent/pkg/version"
 )
@@ -41,6 +39,38 @@ const (
 	packageDatadogInstaller = "datadog-installer"
 	packageAPMLibraryDotnet = "datadog-apm-library-dotnet"
 )
+
+// Installer is a package manager that installs and uninstalls packages.
+type Installer interface {
+	IsInstalled(ctx context.Context, pkg string) (bool, error)
+
+	AvailableDiskSpace() (uint64, error)
+	State(ctx context.Context, pkg string) (repository.State, error)
+	States(ctx context.Context) (map[string]repository.State, error)
+	ConfigState(ctx context.Context, pkg string) (repository.State, error)
+	ConfigStates(ctx context.Context) (map[string]repository.State, error)
+
+	Install(ctx context.Context, url string, args []string) error
+	ForceInstall(ctx context.Context, url string, args []string) error
+	SetupInstaller(ctx context.Context, path string) error
+	Remove(ctx context.Context, pkg string) error
+	Purge(ctx context.Context)
+
+	InstallExperiment(ctx context.Context, url string) error
+	RemoveExperiment(ctx context.Context, pkg string) error
+	PromoteExperiment(ctx context.Context, pkg string) error
+
+	InstallConfigExperiment(ctx context.Context, pkg string, version string, rawConfig []byte) error
+	RemoveConfigExperiment(ctx context.Context, pkg string) error
+	PromoteConfigExperiment(ctx context.Context, pkg string) error
+
+	GarbageCollect(ctx context.Context) error
+
+	InstrumentAPMInjector(ctx context.Context, method string) error
+	UninstrumentAPMInjector(ctx context.Context, method string) error
+
+	Close() error
+}
 
 // installerImpl is the implementation of the package manager.
 type installerImpl struct {
@@ -58,28 +88,12 @@ type installerImpl struct {
 }
 
 // NewInstaller returns a new Package Manager.
-func NewInstaller(env *env.Env) (installertypes.Installer, error) {
+func NewInstaller(env *env.Env) (Installer, error) {
 	err := ensureRepositoriesExist()
 	if err != nil {
 		return nil, fmt.Errorf("could not ensure packages and config directory exists: %w", err)
 	}
-
-	options := []db.Option{
-		db.WithTimeout(10 * time.Second),
-	}
-
-	// On Linux, if we're not root, we can only open the database in read-only mode.
-	if runtime.GOOS == "linux" {
-		currentUser, err := user.Current()
-		if err != nil {
-			return nil, fmt.Errorf("could not get current user: %w", err)
-		}
-		if currentUser.Uid != "0" {
-			options = append(options, db.WithReadOnly())
-		}
-	}
-
-	db, err := db.New(filepath.Join(paths.PackagesPath, "packages.db"), options...)
+	db, err := db.New(filepath.Join(paths.PackagesPath, "packages.db"), db.WithTimeout(10*time.Second))
 	if err != nil {
 		return nil, fmt.Errorf("could not create packages db: %w", err)
 	}
